End to end hooked up
This has sign in / create account, then home with a count for how many
plans current user has, then create plan, then return to list with
updated count. I also added cancels for all modal views
