+++ Quest/ComposerViewController.swift
@@ -14,13 +14,11 @@ class ComposerViewController: UIViewController {
     @IBOutlet weak var titlePlanField: UITextField!
     @IBOutlet weak var tagsPlanField: UITextField!
     @IBOutlet weak var bodyPlanField: UITextField!
-    @IBOutlet weak var totalPlansLabel: UILabel!
     
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        fetchPans()
     }
     
     override func didReceiveMemoryWarning() {
@@ -32,7 +30,6 @@ class ComposerViewController: UIViewController {
     
     @IBAction func didPressCreate(sender: AnyObject) {
         
-//        let user = PFUser.currentUser()
         let plan = PFObject(className: "Plan")
 
         
@@ -46,7 +43,7 @@ class ComposerViewController: UIViewController {
             }
     
         plan["title"] = titlePlanField.text
-        plan["user"] = "testuser"
+        plan["user"] = PFUser.currentUser()?.username
         plan["tags"] = tags
         plan["body"] = bodyPlanField.text
     
@@ -54,10 +51,11 @@ class ComposerViewController: UIViewController {
         plan.saveInBackgroundWithBlock { (status: Bool, error: NSError?) -> Void in
             if error == nil {
                 // print("create success")
-                self.fetchPans()
                 self.titlePlanField.text = ""
                 self.tagsPlanField.text = ""
                 self.bodyPlanField.text = ""
+                self.dismissViewControllerAnimated(true, completion: nil)
+                NSNotificationCenter.defaultCenter().postNotificationName("refresh", object: nil)
                 
             } else {
                 print("error\(error)")
@@ -67,27 +65,10 @@ class ComposerViewController: UIViewController {
         
     }
     
-    // FETCHING NUMBER OF PLANS FROM USER. THIS HAS TESTUSER HARDCODED BUT WE NEED TO CHANGE TO CURRENT USER ONCE EVERYTHING IS CONNECTED. THIS IS A TEST FUNCTION IN THIS SCREEN, WE NEED TO MOVE IT TO THE MAIN LIST VIEW
-    func fetchPans() {
-        
-        let query = PFQuery(className:"Plan")
-        query.whereKey("user", equalTo:"testuser")
-        query.findObjectsInBackgroundWithBlock {
-            (objects: [PFObject]?, error: NSError?) -> Void in
-            
-            if error == nil {
-                // The find succeeded.
-                print("Successfully retrieved \(objects!.count) plans")
-                self.totalPlansLabel.text = String(objects!.count)
-            } else {
-                // Log details of the failure
-                print("Error: \(error!)")
-            }
-        }
-        
-        
+    // CANCELLING PLAN CREATION
+    @IBAction func didPressCancel(sender: AnyObject) {
+        self.dismissViewControllerAnimated(true, completion: nil)
     }
-    
     
 }
