+++ Quest/LoginViewController.swift
@@ -16,17 +16,20 @@ class LoginViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        // Do any additional setup after loading the view.
+
     }
 
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
+
     }
     
+    //LOGIN
     @IBAction func didPressLogIn(sender: AnyObject) {
         PFUser.logInWithUsernameInBackground(userNameField.text!, password: passwordField.text!) { (user: PFUser?, error: NSError?) -> Void in
             if error == nil {
+                self.userNameField.text = ""
+                self.passwordField.text = ""
                 self.performSegueWithIdentifier("logInSegue", sender: self)
             } else {
                 print("error\(error)")
@@ -35,14 +38,10 @@ class LoginViewController: UIViewController {
         
     }
 
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
+    // CANCELLING LOGIN
+    @IBAction func didPressCancel(sender: AnyObject) {
+        self.dismissViewControllerAnimated(true, completion: nil)
     }
-    */
 
+    
 }
